requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
openai==0.28.0
requests==2.31.0
tqdm==4.66.1
yarl==1.9.2
annotated-types==0.6.0
blis==0.7.11
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cloudpathlib==0.15.1
confection==0.1.3
conllu==4.5.3
cymem==2.0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
en-ner-bc5cdr-md @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.3/en_ner_bc5cdr_md-0.5.3.tar.gz#sha256=5fe351539b7a41aa576ae40eb74bb402113ccc1f4b5845d64545f6e2700b1bb3
# en-ner-bionlp13cg-md==0.5.3
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
langcodes==3.3.0
MarkupSafe==2.1.3
murmurhash==1.0.10
nmslib==2.1.1
numpy==1.26.0
packaging==23.2
pathy==0.10.2
preshed==3.0.9
psutil==5.9.5
pybind11==2.6.1
pydantic==2.4.2
pydantic_core==2.10.1
pysbd==0.3.4
requests==2.31.0
scikit-learn==1.3.1
scipy==1.10.1
scispacy==0.5.3
smart-open==6.4.0
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.1.12
threadpoolctl==3.2.0
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
urllib3==2.0.6
wasabi==1.1.2
weasel==0.3.2
Pillow~=10.1.0
neo4j~=5.14.0
flask~=3.0.0
pandas~=2.1.1